pyproject.toml
[tool.poetry]
name = "chatapp"
version = "0.1.0"
description = ""
authors = ["45deg <45deg@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
aiomysql = "^0.2.0"
sqlalchemy = "^2.0.25"
uvicorn = "^0.27.1"
google-cloud-firestore = "^2.14.0"
redis = "^5.0.1"
google-cloud-pubsub = "^2.19.4"

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.19.0"
black = "^24.1.1"
sqlalchemy-stubs = "^0.4"
websockets = "^12.0"
ruff = "^0.2.1"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
